Add delegation util class
